terrain seems to work

+++ market_economy_simulation/src/deferred_terrain_shader/shader.wgsl
@@ -17,7 +17,7 @@ struct InstanceInput {
     @location(5) position: vec3<f32>,
     @location(6) color: vec3<f32>,
     @location(7) entity: vec3<u32>,
-    @location(8) color_heighlights: vec3<u32>,
+    @location(8) color_heighlights: vec3<f32>,
 }
 
 struct VertexOutput {
